// frontend/src/api/model/promptVersionDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * TeamPrompts.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { UserRef } from './userRef';

export interface PromptVersionDto {
  id: string;
  promptId: string;
  /** @nullable */
  parentVersionId: string | null;
  content: string;
  author: UserRef;
  /** @nullable */
  note: string | null;
  isMain: boolean;
  createdAt: string;
}